refactor(tests): pull setting up seasons into helper function
