+++ tests/e2e/specs/in-game/game-over/rankedMatches.spec.js
@@ -1,5 +1,5 @@
 import GameStatus from '../../../../../utils/GameStatus.json';
-import { assertLoss, assertVictory, assertStalemate } from '../../../support/helpers';
+import { assertLoss, assertVictory, assertStalemate, setupSeasons } from '../../../support/helpers';
 import { seasonFixtures } from '../../../fixtures/statsFixtures';
 import { playerOne, playerTwo, playerThree } from '../../../fixtures/userFixtures';
 import { Card } from '../../../fixtures/cards';
@@ -26,49 +26,7 @@ function validateGameResult({ status, winner }, expectedWinner) {
 
 describe('Creating And Updating Ranked Matches', () => {
   beforeEach(function () {
-    cy.wipeDatabase();
-    cy.visit('/');
-
-    // Set up season
-    const [ , diamondsSeason ] = seasonFixtures;
-    diamondsSeason.startTime = dayjs.utc().subtract(2, 'week')
-      .subtract(1, 'day')
-      .toDate();
-    diamondsSeason.endTime = dayjs.utc().add(11, 'weeks')
-      .toDate();
-    cy.loadSeasonFixture([ diamondsSeason ]);
-    // Sign up to players and store their id's for comparison to match data
-    cy.signupOpponent(playerOne).as('playerOneId');
-    cy.signupOpponent(playerThree).as('playerThreeId');
-    // Opponent will be player 2 (the last one we log in as)
-    cy.signupOpponent(playerTwo)
-      .as('playerTwoId')
-      .then(function () {
-        // Create match from last week, which current games don't count towards
-        const oldMatchBetweenPlayers = {
-          player1: this.playerOneId,
-          player2: this.playerTwoId,
-          winner: this.playerOneId,
-          startTime: dayjs.utc().subtract(1, 'week')
-            .subtract(1, 'day')
-            .toDate(),
-          endTime: dayjs.utc().subtract(1, 'week')
-            .subtract(1, 'day')
-            .toDate(),
-        };
-
-        const currentMatchWithDifferentOpponent = {
-          player1: this.playerOneId,
-          player2: this.playerThreeId,
-          winner: null,
-          startTime: dayjs.utc().subtract(1, 'hour')
-            .toDate(),
-          endTime: dayjs.utc().subtract(1, 'hour')
-            .toDate(),
-        };
-
-        cy.loadMatchFixtures([ oldMatchBetweenPlayers, currentMatchWithDifferentOpponent ]);
-      });
+    setupSeasons();
     // Log in as playerOne
     cy.loginPlayer(playerOne);
     cy.setupGameAsP0(true, true);
