+++ tests/e2e/support/helpers.js
@@ -1,5 +1,10 @@
+import dayjs from 'dayjs';
+import utc from 'dayjs/plugin/utc';
 import { SnackBarError } from '../fixtures/snackbarError';
-import { playerOne, playerTwo } from '../fixtures/userFixtures';
+import { playerOne, playerTwo, playerThree } from '../fixtures/userFixtures';
+import { seasonFixtures } from '../fixtures/statsFixtures';
+
+dayjs.extend(utc);
 
 export function hasValidSuitAndRank(card) {
   if (!Object.prototype.hasOwnProperty.call(card, 'rank')) {
@@ -294,6 +299,52 @@ function assertDomMatchesFixture(pNum, fixture, spectating) {
   if (fixture.scrap) {
     cy.get('#scrap').contains(`(${fixture.scrap.length})`);
   }
+}
+
+export function setupSeasons() {
+  cy.wipeDatabase();
+  cy.visit('/');
+
+  // Set up season
+  const [ , diamondsSeason ] = seasonFixtures;
+  diamondsSeason.startTime = dayjs.utc().subtract(2, 'week')
+    .subtract(1, 'day')
+    .toDate();
+  diamondsSeason.endTime = dayjs.utc().add(11, 'weeks')
+    .toDate();
+  cy.loadSeasonFixture([ diamondsSeason ]);
+  // Sign up to players and store their id's for comparison to match data
+  cy.signupOpponent(playerOne).as('playerOneId');
+  cy.signupOpponent(playerThree).as('playerThreeId');
+  // Opponent will be player 2 (the last one we log in as)
+  cy.signupOpponent(playerTwo)
+    .as('playerTwoId')
+    .then(function () {
+      // Create match from last week, which current games don't count towards
+      const oldMatchBetweenPlayers = {
+        player1: this.playerOneId,
+        player2: this.playerTwoId,
+        winner: this.playerOneId,
+        startTime: dayjs.utc().subtract(1, 'week')
+          .subtract(1, 'day')
+          .toDate(),
+        endTime: dayjs.utc().subtract(1, 'week')
+          .subtract(1, 'day')
+          .toDate(),
+      };
+
+      const currentMatchWithDifferentOpponent = {
+        player1: this.playerOneId,
+        player2: this.playerThreeId,
+        winner: null,
+        startTime: dayjs.utc().subtract(1, 'hour')
+          .toDate(),
+        endTime: dayjs.utc().subtract(1, 'hour')
+          .toDate(),
+      };
+
+      cy.loadMatchFixtures([ oldMatchBetweenPlayers, currentMatchWithDifferentOpponent ]);
+    });
 }
 
 /**
